物体识别高于此分数则认为正确 : 0.6

+++ src/java/algorithm-bootstrap/src/main/java/com/sengled/mediaworker/algorithm/ObjectRecognitionImpl.java
@@ -68,7 +68,7 @@ public class ObjectRecognitionImpl implements ObjectRecognition,InitializingBean
     private String debugImageSavePath;
     
     @Value("${object.confirm.score}")
-    private Long objectConfirmScore;
+    private Double objectConfirmScore;
     
     
 	@Autowired
